pymatcher: simple scoring + new regex
The main change here is to the regex. The idea is to be able to search the path
or filename for *all* matching (possibly overlapping) substrings. The search is
also non-greedy. This means the search returns a list of results for we have to
now find the best result to return. I have also simplified the scoring to favor
the shortest results.